server/requirements/base.txt
Django==3.1.1
djangorestframework
markdown
django-filter